Solved the deprecation warning and the starting bugs. Now we have to check the login process.
As I see the code is in between the LeoN to Chalks transition. The freeze seems to had happened just after I figured a smart way to manage the State Vector in a descentralized system (replacing the sites references from a numered list (concept of site_index) to a dictionary of sites ids).

Actually the two nodes startup but nothing happens. Clearly the connection code rewrite is unifinished, will try to check this during the weekend.

(cool, using logs as emails)