Lots of minor edits. Enhanced the connect to dialog. Added some todos. Worked on the Chat. Almost working but finally got a mess with Perspectives, pb.Referenceable and callRemote. Put some comments in the todos, need to sleep (i'm not able to think anymore). Will take back a look after Thursday.
Would be fine to document the Connections schemas (blblablb...Avatar pass Mind that is keeped as a list of perspective that are accessed by parent to send info to childrens...blalblb). I forget easily this kind of things.

If you want you can try to get it work, will help.